broad strokes of attempts refactor, clean compilation
