+++ lib/oli/delivery/attempts/core/client.ex
@@ -1,4 +1,4 @@
-defmodule Oli.Delivery.Attempts.ClientEvaluation do
+defmodule Oli.Delivery.Attempts.Core.ClientEvaluation do
   @enforce_keys [:input, :score, :out_of, :feedback]
   @derive Jason.Encoder
   defstruct [:input, :score, :out_of, :feedback]
