+++ lib/oli/delivery/attempts/core/part_attempt.ex
@@ -1,19 +1,19 @@
-defmodule Oli.Delivery.Attempts.PartAttempt do
+defmodule Oli.Delivery.Attempts.Core.PartAttempt do
   use Ecto.Schema
   import Ecto.Changeset
 
   schema "part_attempts" do
-    field :attempt_guid, :string
-    field :attempt_number, :integer
-    field :date_evaluated, :utc_datetime
-    field :score, :float
-    field :out_of, :float
-    field :response, :map
-    field :feedback, :map
-    field :hints, {:array, :string}, default: []
-    field :part_id, :string
+    field(:attempt_guid, :string)
+    field(:attempt_number, :integer)
+    field(:date_evaluated, :utc_datetime)
+    field(:score, :float)
+    field(:out_of, :float)
+    field(:response, :map)
+    field(:feedback, :map)
+    field(:hints, {:array, :string}, default: [])
+    field(:part_id, :string)
 
-    belongs_to :activity_attempt, Oli.Delivery.Attempts.ActivityAttempt
+    belongs_to(:activity_attempt, Oli.Delivery.Attempts.Core.ActivityAttempt)
 
     timestamps(type: :utc_datetime)
   end
