+++ lib/oli/delivery/attempts/core/resource_attempt.ex
@@ -1,18 +1,18 @@
-defmodule Oli.Delivery.Attempts.ResourceAttempt do
+defmodule Oli.Delivery.Attempts.Core.ResourceAttempt do
   use Ecto.Schema
   import Ecto.Changeset
 
   schema "resource_attempts" do
-    field :attempt_guid, :string
-    field :attempt_number, :integer
-    field :date_evaluated, :utc_datetime
-    field :score, :float
-    field :out_of, :float
-    field :state, :map, default: %{}
+    field(:attempt_guid, :string)
+    field(:attempt_number, :integer)
+    field(:date_evaluated, :utc_datetime)
+    field(:score, :float)
+    field(:out_of, :float)
+    field(:state, :map, default: %{})
 
-    belongs_to :resource_access, Oli.Delivery.Attempts.ResourceAccess
-    belongs_to :revision, Oli.Resources.Revision
-    has_many :activity_attempts, Oli.Delivery.Attempts.ActivityAttempt
+    belongs_to(:resource_access, Oli.Delivery.Attempts.Core.ResourceAccess)
+    belongs_to(:revision, Oli.Resources.Revision)
+    has_many(:activity_attempts, Oli.Delivery.Attempts.Core.ActivityAttempt)
 
     timestamps(type: :utc_datetime)
   end
