+++ lib/oli/delivery/attempts/core/student_input.ex
@@ -1,4 +1,4 @@
-defmodule Oli.Delivery.Attempts.StudentInput do
+defmodule Oli.Delivery.Attempts.Core.StudentInput do
   @enforce_keys [:input]
   @derive Jason.Encoder
   defstruct [:input]
