+++ lib/oli/delivery/evaluation/actions/navigation_action_result.ex
@@ -1,4 +1,4 @@
-defmodule Oli.Delivery.Attempts.NavigationActionResult do
+defmodule Oli.Delivery.Evaluation.Actions.NavigationActionResult do
   @derive Jason.Encoder
   defstruct [:type, :to, :error, :attempt_guid]
 end
