+++ lib/oli/delivery/evaluation/actions/state_update_action_result.ex
@@ -1,4 +1,4 @@
-defmodule Oli.Delivery.Attempts.StateUpdateActionResult do
+defmodule Oli.Delivery.Evaluation.Actions.StateUpdateActionResult do
   @derive Jason.Encoder
   defstruct [:type, :update, :error, :attempt_guid]
 end
