+++ lib/oli/delivery/evaluation/result.ex
@@ -1,4 +1,4 @@
-defmodule Oli.Delivery.Attempts.Result do
+defmodule Oli.Delivery.Evaluation.Result do
   @enforce_keys [:score, :out_of]
   @derive Jason.Encoder
   defstruct [:score, :out_of]
