+++ lib/oli_web/controllers/page_delivery_controller.ex
@@ -7,7 +7,7 @@ defmodule OliWeb.PageDeliveryController do
   alias Oli.Rendering.Context
   alias Oli.Rendering.Page
   alias Oli.Activities
-  alias Oli.Delivery.Attempts
+  alias Oli.Delivery.Attempts.PageLifecycle
   alias Oli.Utils.Slug
   alias Oli.Utils.Time
   alias Oli.Delivery.Sections
@@ -36,7 +36,7 @@ defmodule OliWeb.PageDeliveryController do
     user = conn.assigns.current_user
 
     if Sections.is_enrolled?(user.id, section_slug) do
-      PageContext.create_page_context(section_slug, revision_slug, user)
+      PageContext.create_for_visit(section_slug, revision_slug, user)
       |> render_page(conn, section_slug, user)
     else
       render(conn, "not_authorized.html")
@@ -196,7 +196,7 @@ defmodule OliWeb.PageDeliveryController do
     activity_provider = &Oli.Delivery.ActivityProvider.provide/2
 
     if Sections.is_enrolled?(user.id, section_slug) do
-      case Attempts.start_resource_attempt(
+      case PageLifecycle.start(
              revision_slug,
              section_slug,
              user.id,
@@ -227,14 +227,8 @@ defmodule OliWeb.PageDeliveryController do
     user = conn.assigns.current_user
 
     if Sections.is_enrolled?(user.id, section_slug) do
-      case Attempts.review_resource_attempt(attempt_guid) do
-        {:ok, _} ->
-          PageContext.create_page_context(section_slug, revision_slug, attempt_guid, user)
-          |> render_page(conn, section_slug, user)
-
-        _ ->
-          render(conn, "error.html")
-      end
+      PageContext.create_for_review(section_slug, revision_slug, attempt_guid, user)
+      |> render_page(conn, section_slug, user)
     else
       render(conn, "not_authorized.html")
     end
@@ -265,7 +259,7 @@ defmodule OliWeb.PageDeliveryController do
     user = conn.assigns.current_user
     section = conn.assigns.section
 
-    case Attempts.submit_graded_page(section_slug, attempt_guid) do
+    case PageLifecycle.finalize(section_slug, attempt_guid) do
       {:ok, resource_access} ->
         grade_sync_result = send_one_grade(section, user, resource_access)
         after_finalized(conn, section_slug, revision_slug, attempt_guid, user, grade_sync_result)
@@ -289,7 +283,7 @@ defmodule OliWeb.PageDeliveryController do
   end
 
   def after_finalized(conn, section_slug, revision_slug, attempt_guid, user, grade_sync_result) do
-    context = PageContext.create_page_context(section_slug, revision_slug, user)
+    context = PageContext.create_for_visit(section_slug, revision_slug, user)
 
     message =
       if context.page.max_attempts == 0 do
